+++ client/src/components/ItemDetails/ItemDetails.scss
@@ -1,0 +1,46 @@
+@import '../../styles/variables.scss';
+
+.item-details {
+  display: grid;
+  background-color: $col-bright;
+  outline: 2px solid transparent;
+  box-shadow: 1px 1px 20px $shadow-500;
+  border-radius: 1rem;
+
+  // grid row, with grid-auto-rows in parent define the subgrid
+  grid-row: span 3;
+  grid-template-rows: subgrid;
+  gap: 0.5rem;
+
+  &__name {
+    font: $text-h2;
+    margin: 0;
+  }
+
+  &__descr {
+    font: $text-normal;
+  }
+
+  &__img {
+    object-fit: cover;
+    border-radius: 0.9rem 0.9rem 0 0;
+    width: 100%;
+    height: 152px;
+    // border-radius: 12px;
+  }
+  &__link {
+    display: grid;
+    grid-template-rows: subgrid;
+    grid-row: span 3;
+    text-decoration: none;
+    color: inherit;
+  }
+
+  :not(img) {
+    padding: 0 0.5rem;
+  }
+
+  &:hover {
+    outline: 2px solid $col-prim-500;
+  }
+}
