+++ client/src/pages/DetailsPage/DetailsPage.tsx
@@ -1,27 +1,22 @@
-import {useState, useEffect} from 'react';
 import {useParams} from 'react-router-dom';
-import {fetchItems} from '../../services/ItemServices';
-import {logger} from '../../util/logger';
-import catchAsync from '../../util/catchAsync';
+
 import NoAuthRedirect from '../../components/NoAuthRedirect';
+import ItemDetails from '../../components/ItemDetails/ItemDetails';
+import Button from '../../components/Button/Button';
+import Chat from '../../components/Chat/Chat';
 
 function DetailsPage(): JSX.Element {
-  const {id} = useParams();
-  const [item, setItem] = useState({});
-
-  useEffect(() => {
-    logger.log('useEffect started');
-    const fetchData = catchAsync(async () => {
-      const fetchedItem = await fetchItems(`books/${id}`);
-      setItem(fetchedItem.data);
-      logger.log('DetailsPage fetched this item: ', fetchedItem);
-    });
-    fetchData();
-  }, [id]);
-
   // NoAuthRedirect();
 
-  return <div>Details Page</div>;
+  const {id} = useParams();
+
+  return (
+    <div>
+      <ItemDetails id={id} />
+      <Chat />
+      <Button>Reqest Item</Button>
+    </div>
+  );
 }
 
 export default DetailsPage;
